Sublist on Elixir: refactor with some ideas from other users
Looking at other users' solutions, I realize I don't have to *loop* over all
elements in the lists to know if they are equal, it's enough to use

    a === b

or even pattern matching

    def compare(a, a), do: :equal

Other users use `Stream.chunk` to do what I did with several methods and
recursion. It's much easier with `chunk`, it gives you a list of lists, each
one with the length you specified. Then, one just has to compare with `===` to
know if a sublist is equal to a list. Impresive!