Triangle on Elixir: refactor to lear how to use `with/1`
[`with/1`](http://elixir-lang.org/docs/stable/elixir/Kernel.SpecialForms.html#with/1)
is used to combine several matching clauses.

It's great to chain several validation methods because I can remove
one body in some methods (the one that matches the `:error` case) and
I don't need to carry those `:error` or `:ok` cases in the parameters.

On the other side, the solution doesn't make use of the pipe `|>`
operator, but I see that methods are simpler.

I still use a `case` statement on validation methods to return
`:ok` or `:error`, and it still annoys me a little bit.