Leap on Elixir: solve the exercise with a different approach
The previous oslution was nice, but I'd like to try another approach.
In this one, the leap year is divided always by the three posibilities,
creating a matrix. Then, with pattern matching, the right return value
is selected.

It's a bit obscure and cryptic, but it's different, and I like it because
of it.